Allow adding embed URLs by drag and dropping
